// frontend/src/lib/types/profile.ts
export interface Profile {
  followers_count: string;
  following_count: string;
  is_verified: string;
  posts_count: string;
  posts: string;
  bio: string;
  name: string;
  profilePicture: string;
  platform: string;
  profile_url: string;
  page_name: string;
  page_likes: string;
  page_followers: string;
  page_category: string;
  page_website: string;
}